requirements.txt
langchain_community
langchain_core
langchain_text_splitters
langchain_google_genai
streamlit
pypdf
python-docx
pinecone
python-dotenv
google-generativeai